+++ extensions/reclaim-ai/CHANGELOG.md
@@ -1,5 +1,8 @@
 # reclaim Changelog
 
+## [Fixes] - {PR_MERGE_DATE}
+- Don't show declined events in the menu bar title or My Calendar when the user has "show declined events" disabled
+
 ## [Update] - 2025-10-30
 - Update to the package.json platforms property to include Windows.
 - Update to Search Tasks to not include tasks that are cancelled.